+++ src/Services/NotifierDatabaseService.php
@@ -6,7 +6,6 @@ use Devuni\Notifier\Support\NotifierLogger;
 use Throwable;
 use Carbon\Carbon;
 use GuzzleHttp\Client;
-use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\File;
 use Symfony\Component\Process\Process;
 
@@ -76,7 +75,7 @@ class NotifierDatabaseService
             }
         } catch (Throwable $th) {
             NotifierLogger::get()->emergency('❌ an error occurred while uploading a file', [
-                'th' => $th->getMessage(),
+                'error' => $th->getMessage(),
                 'env' => config('notifier.backup_url'),
                 'code' => config('notifier.backup_code'),
             ]);